Creates testing web server.
In order to run a test web server we import the [Microsoft.AspNetCore.Mvc.Testing](https://www.nuget.org/packages/Microsoft.AspNetCore.Mvc.Testing). We override the `ConfigureWebHost` method to [customize]((https://docs.microsoft.com/en-us/aspnet/core/test/integration-tests?view=aspnetcore-5.0#customize-webapplicationfactory) our IOC container with any testing dependencies we may need. In this case we have set it to use a test database.

By default `WebApplicationFactory` is configured to contain test execution within the test process. The webserver doesn't open a network port and communication must be done with `HttpClient` the `WebApplicationFactory` has configured for in process communication. This won't allow an external process like a web browser to connect to the web server.

We can override this behavior by replacing `CreateServer` with an implementation that starts a web server that will host an external port. For web projects using `IWebHostBuilder` this is all that is required. This was the default for the ASP.Net core 2.x project templates. For ASP.Net core 5.0 project templates were switched to use `IHostBuilder`. For our test server to work we need to override `CreateWebHostBuilder` with an implementation that mimics the one called in `Program.Main` of our web project.

This was based on Scot Hanselman's [Real Browser Integration Testing with Selenium Standalone, Chrome, and ASP.NET Core 2.1](https://www.hanselman.com/blog/real-browser-integration-testing-with-selenium-standalone-chrome-and-aspnet-core-21).