Updating Startup.Startup
Since we called the testing environment `IntegrationTest` we need to make some changes to `Startup` to ensure the web app will run.

To provide us with better error messages we enable the developer exception page.

The test application is in a different folder. To ensure the angular application is built and started configure the source path and use the AngularCLI server.