Making XPath queries more resilient.
The XPath queries we used to select elements in the previous commit are extremely fragile. For example re-ordering the options in the nav menu would cause the test to fail because the Counter link would no longer be the second item in the list (`li[2]`). We want to identify less volatile identifiers for the items we interact with.

Similar to how WPF applications are divided into controls, Angular applications are divided into components. When a component is used Angular inserts custom HTML tags for that component that wrap its contents. In our example we have components for the nav-menu and the counter page. By updating the XPath to look for the components anywhere on the page we are isolated from changes to the structure of the page.