Page object model.
It is easier to deal with changes in our application if we isolate those changes. For this we can use (page object models)[https://www.selenium.dev/documentation/en/guidelines_and_recommendations/page_object_models/]. In the page object model design pattern we create a class for each page. These page models contain the code for interacting with the page. If the action navigates to a new page then a new page object model is returned.