Page object model continued.
We want to give the same treatment to our count page. We make Increment a method it is an action, but we don't return a new page because it doesn't' navigate. CurrentCount is a property because it is a value on the page. We can even go one step further and parse the value to an integer.

Notice how clear the end of our test is becoming. We open the count page, increment the count, and check the value.