Reusing test setup.
In XUnit shared test stup is done in the constructor. We can introduce a base class to handle setting up and tearing down the test environment. Fixtures can be used to share items with expensive startup/shutdown times, but care needs to be taken to ensure the environment is properly reset between tests.