Testing adding data
We are continuing to create classes to encapsulate child elements. This is becoming more important as the test grows in complexity.

Since starting a new test is expensive we will continue where our old test left off.

We can leverage both the page objects and our database context to verify the application is working. While the page's current implementation forces a refresh we don't want to risk the front end tricking our test into thinking everything is fine if the API call failed. To give the perception of better performance some applications will simulate changes on the front end and roll them back on failure. An alternative could be to navigate away from the page.